Reintroduce the EventContext class
Another substantial refactoring. Work on the event handler instrumentation
revealed the limitation of the current approach - a single thread-specific
EventRacer log simply does not cut it, because we need multiple active logs at
the same time. For example, cancellation of requests or unload events, occurring
during committing a provisional load go in a different log than the one of the
said provisional load. This changeset does not yet solve the issue, but should
provide mechanisms for solving it.