Delay creation of the EventRacer host part until URL load is comitted
When we create an EventRacer log at the host, the previous log, beloging to
the same site instance, is written to disk and stops receiving messages. When
we create the host log too early (as we did), but don't commit to the
navigation we miss the rest of the action.