Ensure image load/error events are sent before frame completion check
If |Document::implicitClose| called |ImageLoader::dispatchPendingLoadEvents|
and in fact there are load events pending (in the |EventSender|) then
the execution of the event handlers will be contained in a single
event-action, whereas they should be concurrent. To make sure no
events are pending in this case, we use the document's load event delay
counter/timer mechanism to postpone the document load while there are
events queued in the |ImageLoader|'s load and error |EventSender|s.
Now the calls to |ImageLoader::dispatchPendingLoadEvents| and
|ImageLoader::dispatchPendingErrorEvents| from |Document::implicitClose|
ought not to be needed. They are kept around for a little bit of extra safety.